requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.0.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.1.5
Flask-WTF==0.14.2
gcloud==0.17.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
httplib2==0.15.0
itsdangerous==1.1.0
Jinja2==2.10.3
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
passlib==1.7.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dotenv==0.10.3
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
WTForms==2.2.1
